# DTC.Api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DTC.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/matches/round/{roundId}": {
      "get": {
        "tags": [
          "Match"
        ],
        "parameters": [
          {
            "name": "roundId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{id}": {
      "get": {
        "tags": [
          "Match"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Match"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Match"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/matches": {
      "post": {
        "tags": [
          "Match"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/matchparticipants/match/{matchId}": {
      "get": {
        "tags": [
          "MatchParticipant"
        ],
        "parameters": [
          {
            "name": "matchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchParticipantDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchParticipantDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchParticipantDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rounds/tournament/{tournamentId}": {
      "get": {
        "tags": [
          "Braket"
        ],
        "parameters": [
          {
            "name": "tournamentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BraketDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BraketDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BraketDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tournaments": {
      "get": {
        "tags": [
          "Tournament"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tournament"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{id}": {
      "get": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTournamentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/tournaments/{id}/generate-groups": {
      "post": {
        "tags": [
          "Tournament"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateGroupsDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateGroupsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateGroupsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateGroupsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateMatchDto": {
        "required": [
          "roundId"
        ],
        "type": "object",
        "properties": {
          "roundId": {
            "type": "integer",
            "format": "int32"
          },
          "boardId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          }
        },
        "additionalProperties": false
      },
...
      "CreateTournamentDto": {
        "required": [
          "name",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/TournamentMode"
          },
          "status": {
            "$ref": "#/components/schemas/TournamentStatus"
          }
        },
        "additionalProperties": false
      },
      "GenerateGroupsDto": {
        "type": "object",
        "properties": {
          "groupSize": {
            "type": "integer",
            "format": "int32"
          },
          "playerIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MatchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roundId": {
            "type": "integer",
            "format": "int32"
          },
          "boardId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "actualStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchParticipantDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MatchParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "matchId": {
            "type": "integer",
            "format": "int32"
          },
          "playerId": {
            "type": "integer",
            "format": "int32"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerDto"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "isWinner": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "MatchStatus": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "PlayerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "BraketDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "tournamentId": {
            "type": "integer",
            "format": "int32"
          },
          "locationId": {
            "type": "integer",
            "format": "int32"
          },
          "sequence": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "plannedStart": {
            "type": "string",
            "format": "date-time"
          },
          "plannedEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/BraketStatus"
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BraketStatus": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "TournamentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/TournamentMode"
          },
          "status": {
            "$ref": "#/components/schemas/TournamentStatus"
          },
          "rounds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BraketDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TournamentMode": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      },
      "TournamentStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "UpdateMatchDto": {
        "type": "object",
        "properties": {
          "boardId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "actualStart": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "actualEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTournamentDto": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "mode": {
            "$ref": "#/components/schemas/TournamentMode"
          },
          "status": {
            "$ref": "#/components/schemas/TournamentStatus"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}